🏗️ Reorganize repository structure to professional standards
- Remove screenshots section from README for cleaner presentation
- Move components and lib to src/ directory following standard conventions
- Consolidate all configuration files in config/ directory
- Organize data files and scripts into dedicated data/ and tools/ directories
- Rename tests to __tests__ following Jest conventions
- Update package.json scripts and import paths for new structure
- Create symlinks for config files that must remain in root
- Maintain TypeScript path mapping compatibility

Repository now follows enterprise-level organization with clear separation of concerns.
