+++ config/jest.config.js
@@ -2,7 +2,7 @@ const nextJest = require('next/jest')
 
 const createJestConfig = nextJest({
   // Provide the path to your Next.js app to load next.config.js and .env files
-  dir: './',
+  dir: '../',
 })
 
 // Add any custom config to be passed to Jest
@@ -14,22 +14,20 @@ const customJestConfig = {
   },
   testEnvironment: 'jest-environment-jsdom', // Use jsdom for React components
   testMatch: [
-    '<rootDir>/tests/**/*.(test|spec).(ts|tsx|js)',
+    '<rootDir>/__tests__/**/*.(test|spec).(ts|tsx|js)',
     '<rootDir>/app/**/__tests__/**/*.(test|spec).(ts|tsx|js)',
-    '<rootDir>/components/**/__tests__/**/*.(test|spec).(ts|tsx|js)',
-    '<rootDir>/lib/**/__tests__/**/*.(test|spec).(ts|tsx|js)'
+    '<rootDir>/src/**/__tests__/**/*.(test|spec).(ts|tsx|js)'
   ],
   collectCoverageFrom: [
     'app/**/*.{js,jsx,ts,tsx}',
-    'components/**/*.{js,jsx,ts,tsx}',
-    'lib/**/*.{js,jsx,ts,tsx}',
+    'src/**/*.{js,jsx,ts,tsx}',
     '!**/*.d.ts',
     '!**/layout.tsx',
     '!**/globals.css',
     '!**/__tests__/**',
     '!**/*.test.*',
     '!**/*.spec.*',
-    '!tests/**/*'
+    '!__tests__/**/*'
   ],
   coverageThreshold: {
     global: {
