+++ package.json
@@ -9,18 +9,18 @@
     "lint": "next lint",
     "lint:fix": "next lint --fix",
     "typecheck": "tsc --noEmit",
-    "test": "jest",
-    "test:unit": "jest tests/unit",
-    "test:integration": "jest tests/integration",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
-    "test:ci": "jest --ci --coverage --watchAll=false",
-    "db:test": "node scripts/test-supabase-connection.js",
-    "db:setup": "node scripts/setup-supabase-schema.js",
-    "db:status": "node scripts/check-supabase-status.js",
+    "test": "jest --config=config/jest.config.js",
+    "test:unit": "jest __tests__/unit --config=config/jest.config.js",
+    "test:integration": "jest __tests__/integration --config=config/jest.config.js",
+    "test:watch": "jest --watch --config=config/jest.config.js",
+    "test:coverage": "jest --coverage --config=config/jest.config.js",
+    "test:ci": "jest --ci --coverage --watchAll=false --config=config/jest.config.js",
+    "db:test": "node tools/deployment/test-supabase-connection.js",
+    "db:setup": "node tools/deployment/setup-supabase-schema.js",
+    "db:status": "node tools/deployment/check-supabase-status.js",
     "db:migrate": "psql $DATABASE_URL -f config/database/001-core-schema.sql -f config/database/002-materialized-views.sql -f config/database/003-rls-policies.sql",
-    "env:check": "node scripts/validate-env.js",
-    "screenshots": "node scripts/generate-screenshots.js",
+    "env:check": "node tools/deployment/validate-env.js",
+    "screenshots": "node tools/deployment/generate-screenshots.js",
     "docs:images": "npm run screenshots",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
