+++ src/lib/insight-engine.ts
