+++ tailwind.config.js
@@ -1,34 +1,1 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    './pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './components/**/*.{js,ts,jsx,tsx,mdx}',
-    './app/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/**/*.{js,ts,jsx,tsx,mdx}',
-  ],
-  theme: {
-    extend: {
-      colors: {
-        primary: {
-          50: '#eff6ff',
-          500: '#3b82f6',
-          600: '#2563eb',
-          700: '#1d4ed8',
-        },
-        success: {
-          50: '#f0fdf4',
-          500: '#22c55e',
-        },
-        warning: {
-          50: '#fffbeb',
-          500: '#f59e0b',
-        },
-        error: {
-          50: '#fef2f2',
-          500: '#ef4444',
-        }
-      },
-    },
-  },
-  plugins: [],
-}
+config/tailwind.config.js
