+++ tools/deployment/generate-screenshots.js
